+++ ms-reservas/src/main/java/com/eventos/ms_reservas/exception/FechaInvalidaException.java
@@ -5,14 +5,14 @@ package com.eventos.ms_reservas.exception;
  */
 public class FechaInvalidaException extends RuntimeException {
 
-    private final Long id;
+    private final Integer id;
 
-    public FechaInvalidaException(Long id, String message) {
+    public FechaInvalidaException(Integer id, String message) {
         super(message);
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
